Adding email verification

+++ backend/urban_waste_recycle/src/main/java/com/example/urban_watse_recycling/urban_waste_recycle/Service/UserService.java
@@ -1,12 +1,11 @@
 package com.example.urban_watse_recycling.urban_waste_recycle.Service;
 
 import com.example.urban_watse_recycling.urban_waste_recycle.Model.User;
-
 import java.util.List;
 
 public interface UserService {
     String add(User user);
-    User getById(int id);
+    User getById(Long id);
     List<User> getAll();
     String login(User user);
 }
